Move benefit_estimate to its own folder
